Measure the execution time of Simulations (#11)
* added an assertion to report a wrong input dimension

* Implemented a class to measure the execution time

* print the execution time of simulations

+++ src/test/java/org/group_mmm/ArgParserTest.java
@@ -482,7 +482,7 @@ class ArgParserTest {
                 assertEquals("output.mapper.txt", argParser.getOutputMapperFile());
                 assertEquals(ArgParser.EquivType.SA, argParser.getEquiv());
                 assertEquals("stl.txt", argParser.getStlFile());
-                assertEquals(0.2, argParser.getAlpha());
+                assertEquals(0.2, (double) argParser.getAlpha());
                 assertNull(argParser.getStlFormula());
             }
